Move sunburst helpers to hierarchy.ts module
